UI/requirements.txt
selenium>=4.12.0
pytest>=8.0.0
webdriver-manager>=3.8.6
python-dotenv>=1.0.0
Faker